# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This api is the backend api responsible for the whole Coo-e coordinator app.",
    "title": "Coordinator",
    "version": "",
    "x-stoplight": {
      "id": "haridasu/specs/coordinator"
    }
  },
  "host": "resources.coo-e.com:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/teeups/auto/happeningIfWhenIsReached/{teeup_id}": {
      "post": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "happeningIfWhenIsReached": true,
                "message": "Success",
                "status": 200
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Invalid field parameters, takes boolean fields isPublic, join, invite, suggest, decideGamePlan, and modifyRow.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "",
        "summary": "Happening if when is reached auto change"
      }
    },
    "/api/teeups/auto/itsOnMin/{teeup_id}": {
      "post": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "message": "Success",
                "newMin": 2,
                "status": 200
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Invalid field parameters, takes boolean fields isPublic, join, invite, suggest, decideGamePlan, and modifyRow.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "",
        "summary": "Its on min"
      }
    },
    "/api/teeups/auto/requireItsOnBeSetFirst/{teeup_id}": {
      "post": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "message": "Success",
                "requireItsOnBeSetFirst": true,
                "status": 200
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Invalid field parameters, takes boolean fields isPublic, join, invite, suggest, decideGamePlan, and modifyRow.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "",
        "summary": "Require It's On Be Set First for auto change"
      }
    },
    "/api/teeups/auto/requireWhenBeDecided/{teeup_id}": {
      "post": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "message": "Success",
                "requireWhenBeDecided": true,
                "status": 200
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Invalid field parameters, takes boolean fields isPublic, join, invite, suggest, decideGamePlan, and modifyRow.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "",
        "summary": "Require when be decided for auto change"
      }
    },
...
  },
...
}